refactored rest of code
